FIX: DefaultLegacyOpener::open() doesn't search plugins on null input
Perhaps because of a bug elsewhere, there exists a case when pressing
ImageJ/Fiji GUI -> File -> Open triggers the DefaultLegacyOpener
even before the file open dialog comes out. In this case, the open()
is called with 'null' parameter (as nothing has yet been chosen/selected
for opening), leads into the call getEagerHander() and breaks there.
The fix simply allows to start looking for potential "eager" plugin
only when the input is valid (that is, not null).